requirements.txt
Django==4.0.4
djangorestframework==3.13.1
drf-spectacular==0.22.1